DFE-887 Extract caption into DataTableCaption component
