--- src/explore-education-statistics-frontend/src/modules/table-tool/components/DataTableKeys.tsx
@@ -1,16 +1,0 @@
-import React from 'react';
-
-const DataTableKeys = () => {
-  return (
-    <ul>
-      <li>
-        <strong>n/a</strong> - value matching this criteria could not be found.
-      </li>
-      <li>
-        <strong>x</strong> - value matching this criteria is suppressed.
-      </li>
-    </ul>
-  );
-};
-
-export default DataTableKeys;
